Word Embeddings for Deep Learning
The web is loaded with textual content, and Natural
Language Processing is a standout amongst the most
vital elds in Machine Learning. But when data is
huge simple Machine Learning algorithms are not able
to handle it and that is when Deep Learning comes
into play which based on Neural Networks. However
since neural networks cannot process raw text, we have
to change over them through some diverse strategies of
word embedding. This paper demonstrates those dis-
tinctive word embedding strategies implemented on an
Amazon Review Dataset, which has two sentiments to
be classied: Happy and Unhappy based on numerous
customer reviews. Moreover we demonstrate the dis-
tinction in accuracy with a discourse about which word
embedding to apply when.